test(node): Merge `NpmVersionUrlFunTest` into `NpmFunTest`
The version URL test case is similar to the test cases in `NpmFunTest`,
so move it there for consistency.
